requirements.txt
hydra-core == 1.3.2
omegaconf == 2.3.0
wandb == 0.17.8
dvc == 3.55.0
scikit-learn == 1.5.1
numpy == 2.1.0
pandas == 2.2.2
rdkit == 2024.3.5
scipy == 1.14.1
matplotlib == 3.9.2
seaborn == 0.13.2
pytorch-lightning == 2.4.0
torch_geometric == 2.5.3
torch == 2.4.0
pyopenssl == 24.1.0
